backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Google Cloud & Gemini
google-generativeai==0.3.2
google-cloud-aiplatform==1.42.1
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.16.4
google-adk==0.1.0  # Agent Development Kit

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1
python-multipart==0.0.6

# Data Processing
pillow==10.2.0
python-magic==0.4.27
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyyaml==6.0.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Monitoring & Logging
structlog==24.1.0
python-json-logger==2.0.7
